requirements.txt
python-telegram-bot==20.3
mistralai
pytube
requests